Beginning to form together

+++ Eventory/Eventory/Profile.swift
@@ -63,7 +63,9 @@ class Profile : NSManagedObject {
         let fetchRequest = NSFetchRequest(entityName: "Profile")
         let fetchResults = (ctx!.executeFetchRequest(fetchRequest, error: nil) as? [Profile])!
         for (var i = 0; i < fetchResults.count ; i++) {
+            if (fetchResults[i].profid != FBSDKAccessToken.currentAccessToken().userID) {
             ctx?.deleteObject(fetchResults[i]);
+            }
         }
         ctx?.save(nil)
     }
